% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data document file.R, R/data-docs.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Dataset}
\format{
A data frame with 5 rows and 2 columns:
\describe{
  \item{x}{The treatment / exposure variable}
  \item{x.d}{A matrix containing exposure and all covariates variables}
  \item{coef.z}{Coefficients for defining zero-inflation in ZI mediator}
  \item{coef.d}{Coefficients for defining non-zero inflated class distribution}
  \item{mdist}{The distribution of mediator}
  \item{theta}{The dispersion parameter for mediator}
  \item{ydist}{The distribution of outcome}
  \item{coef.y}{Coefficients of outcome model}
}

A data frame with N rows and M variables:
\describe{
  \item{var1}{Description of var1}
  \item{var2}{Description of var2}
  ...
}
}
\source{
Your name or source info
}
\usage{
data(example_data)

example_data
}
\description{
This dataset is provided as an example for testing the package functions.

A short description of your dataset.
}
\examples{
data(example_data)
head(example_data)
}
\keyword{datasets}
