% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medbayes.R
\name{medbayes}
\alias{medbayes}
\title{Run Bayesian mediation model}
\usage{
medbayes(
  model.m = hnb.m,
  model.y = hnb.y,
  treat = "treatment",
  mediator = "mediator",
  ind_mediator = NULL,
  outcome = "outcome",
  control.value = 0,
  treat.value = 1,
  logM = FALSE
)
}
\arguments{
\item{model.m}{A brms model for the mediator}

\item{dat.new}{New data for prediction}
}
\value{
Posterior predictions (first column)
}
\description{
Run Bayesian mediation model
}
