\name{mediationBayes}
\Rdversion{1.1.0}
\alias{mediationBayes-package}
\alias{mediationBayes}
\docType{package}
\title{
  Bayesian mediation analysis for zero-inflated data.
}

\description{
  This R package provides a flexible Bayesian mediation analysis framework particularly suitable for zero-inflated data.
  Both zero-inflated mediators and zero-inflated outcomes are handled by this package,
  compatible with a wide range of outcome/mediator distributions.
  This novel technique employs Bayesian models for both the mediator and the outcome,
  utilizing Markov Chain Monte Carlo (MCMC) algorithms for parameter estimation.
  This method can be used to conduct mediation analysis with not only zero-inflated mediators
  but also a wide range of mediator and outcome distributions that can be fitted with GLMs.
}

\details{
  \tabular{ll}{
    Package: \tab mediationBayes\cr
    Type: \tab Package\cr
    Version: \tab 1.1.0\cr
    Date: \tab 2024-11-11\cr
    License: \tab MIT\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Jinhong Cui, jhcui@uab.edu
}
