% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medbayes.R
\docType{package}
\name{mediationBayes}
\alias{mediationBayes}
\alias{medbayes}
\title{mediationBayes Package Main Function}
\usage{
medbayes(
  model.m = hnb.m,
  model.y = hnb.y,
  treat = "treatment",
  mediator = "mediator",
  ind_mediator = NULL,
  outcome = "outcome",
  control.value = 0,
  treat.value = 1,
  logM = FALSE
)
}
